backend/requirements.txt
bcrypt==4.1.2
Flask==3.0.2
Flask-API @ git+https://github.com/flask-api/flask-api@5c92f76
Flask-Cors==4.0.0
jsons>=1.6.3
pycryptodome>=3.20.0
pytest>=8.0.2
